[Refactor] Remove QrCodeFragment and put the logic in ReceiveViewModel Remove ActivityExt, move startQrCodeScan to ActivityHelper
