+++ app/src/test/java/com/jim/sharetocomputer/ui/receive/ReceiveNavigationTest.kt
@@ -20,43 +20,25 @@ package com.jim.sharetocomputer.ui.receive
 
 import android.app.Application
 import androidx.fragment.app.testing.launchFragment
-import androidx.navigation.NavController
-import androidx.navigation.Navigation
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.espresso.intent.matcher.IntentMatchers
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.jim.sharetocomputer.WebUploadService
-import com.jim.sharetocomputer.ui.main.MainFragmentDirections
+import com.jim.sharetocomputer.gateway.ActivityHelper
 import com.jim.sharetocomputer.ui.setting.SettingFragment
 import org.junit.Assert
 import org.junit.Test
 import org.junit.runner.RunWith
-import org.mockito.Mockito
 import org.robolectric.Shadows
 
 @RunWith(AndroidJUnit4::class)
 class ReceiveNavigationTest {
 
     @Test
-    fun open_scan_qr_code_fragment() {
-        val fragmentScenario = launchFragment<SettingFragment>()
-        fragmentScenario.onFragment { fragment ->
-            val navigation = ReceiveNavigation(fragment)
-            val controller = Mockito.mock(NavController::class.java)
-
-            Navigation.setViewNavController(fragment.requireView(), controller)
-
-            navigation.openScanQrCode()
-
-            Mockito.verify(controller).navigate(MainFragmentDirections.actionFragmentMainToFragmentQrcode())
-        }
-    }
-
-    @Test
     fun start_web_upload_service() {
         val fragmentScenario = launchFragment<SettingFragment>()
         fragmentScenario.onFragment { fragment ->
-            val navigation = ReceiveNavigation(fragment)
+            val navigation = ReceiveNavigation(fragment, ActivityHelper())
 
             navigation.startWebUploadService()
